Add ENABLE_SD support to tests Makefile and fix format warnings
- Add ENABLE_SD option to tests Makefile for building with SD card support
- Pass ENABLE_SD to HAL and Hive library builds
- Use Makefile.stm32 for correct STM32 target compilation
- Add BOARD_MOUNTS_SRC for Crazyflie-specific mount table
- Fix format warnings in test_sd.c and test_flash.c (uint32_t needs %lu
  on ARM where it's unsigned long)

Build SD test with:
  make PLATFORM=crazyflie TEST=sd ENABLE_SD=1 crazyflie

+++ examples/pilot/tests/test_flash.c
@@ -142,7 +142,8 @@ int test_flash_run(bool standalone) {
     }
 
     uint32_t erase_time = hal_get_time_ms() - start_time;
-    HIVE_LOG_INFO("Flash erased and opened (fd=%d, %u ms)", fd, erase_time);
+    HIVE_LOG_INFO("Flash erased and opened (fd=%d, %lu ms)", fd,
+                  (unsigned long)erase_time);
 
     // Write test pattern in blocks
     HIVE_LOG_INFO("Writing %d bytes in %d-byte blocks...", TEST_PATTERN_SIZE,
@@ -175,7 +176,7 @@ int test_flash_run(bool standalone) {
     (void)total_written; // Suppress unused warning
 
     uint32_t write_time = hal_get_time_ms() - start_time;
-    HIVE_LOG_INFO("Write OK (%u ms)", write_time);
+    HIVE_LOG_INFO("Write OK (%lu ms)", (unsigned long)write_time);
 
     // Sync and close
     HIVE_LOG_INFO("Syncing to flash...");
@@ -191,7 +192,7 @@ int test_flash_run(bool standalone) {
         }
         return -1;
     }
-    HIVE_LOG_INFO("Sync OK (%u ms)", sync_time);
+    HIVE_LOG_INFO("Sync OK (%lu ms)", (unsigned long)sync_time);
 
     status = hive_file_close(fd);
     if (HIVE_FAILED(status)) {
@@ -276,7 +277,7 @@ int test_flash_run(bool standalone) {
         }
         return -1;
     }
-    HIVE_LOG_INFO("Read and verify OK (%u ms)", read_time);
+    HIVE_LOG_INFO("Read and verify OK (%lu ms)", (unsigned long)read_time);
 
     hive_file_close(fd);
 
@@ -289,12 +290,17 @@ int test_flash_run(bool standalone) {
     HIVE_LOG_INFO("========================================");
     HIVE_LOG_INFO("  Path:       /log");
     HIVE_LOG_INFO("  Size:       %d bytes", TEST_PATTERN_SIZE);
-    HIVE_LOG_INFO("  Erase time: %u ms", erase_time);
-    HIVE_LOG_INFO("  Write time: %u ms (%u bytes/sec)", write_time,
-                  write_time > 0 ? (TEST_PATTERN_SIZE * 1000) / write_time : 0);
-    HIVE_LOG_INFO("  Sync time:  %u ms", sync_time);
-    HIVE_LOG_INFO("  Read time:  %u ms (%u bytes/sec)", read_time,
-                  read_time > 0 ? (TEST_PATTERN_SIZE * 1000) / read_time : 0);
+    HIVE_LOG_INFO("  Erase time: %lu ms", (unsigned long)erase_time);
+    HIVE_LOG_INFO("  Write time: %lu ms (%lu bytes/sec)",
+                  (unsigned long)write_time,
+                  write_time > 0
+                      ? (unsigned long)((TEST_PATTERN_SIZE * 1000) / write_time)
+                      : 0UL);
+    HIVE_LOG_INFO("  Sync time:  %lu ms", (unsigned long)sync_time);
+    HIVE_LOG_INFO(
+        "  Read time:  %lu ms (%lu bytes/sec)", (unsigned long)read_time,
+        read_time > 0 ? (unsigned long)((TEST_PATTERN_SIZE * 1000) / read_time)
+                      : 0UL);
 
     if (standalone) {
         hive_cleanup();
